+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/tool/_internal/InstructionsOverviewTool.kt
@@ -1,15 +1,23 @@
 package dev.turingcomplete.intellijbytecodeplugin.tool._internal
 
+import com.intellij.openapi.actionSystem.DataProvider
+import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper.IdeModalityType
 import com.intellij.ui.ScrollPaneFactory
 import com.intellij.ui.TableSpeedSearch
 import com.intellij.ui.table.JBTable
+import com.intellij.util.ui.GridBag
 import com.intellij.util.ui.UIUtil
+import dev.turingcomplete.intellijbytecodeplugin._ui.CopyValueAction
 import dev.turingcomplete.intellijbytecodeplugin._ui.UiUtils
+import dev.turingcomplete.intellijbytecodeplugin._ui.UiUtils.Table.getSingleSelectedValue
+import dev.turingcomplete.intellijbytecodeplugin._ui.ViewValueAction
 import dev.turingcomplete.intellijbytecodeplugin._ui.getMaxRowWith
 import dev.turingcomplete.intellijbytecodeplugin._ui.overrideLeftInset
 import dev.turingcomplete.intellijbytecodeplugin._ui.overrideTopInset
+import dev.turingcomplete.intellijbytecodeplugin._ui.withCommonsDefaults
+import dev.turingcomplete.intellijbytecodeplugin.common.CommonDataKeys
 import dev.turingcomplete.intellijbytecodeplugin.tool.ByteCodeTool
 import java.awt.Component
 import java.awt.Dimension
@@ -45,20 +53,23 @@ class InstructionsOverviewTool : ByteCodeTool("Instructions Overview") {
     }
 
     UiUtils.Dialog.show("Instructions Overview", JPanel(GridBagLayout()).apply {
-      val bag = UiUtils.createDefaultGridBag().setDefaultAnchor(GridBagConstraints.WEST)
+      val bag = GridBag().withCommonsDefaults().setDefaultAnchor(GridBagConstraints.WEST)
 
       val table = ScrollPaneFactory.createScrollPane(InstructionsTable())
       add(table, bag.nextLine().next().coverLine().fillCell().weightx(1.0).weighty(1.0))
 
       add(JLabel("Source:"), bag.nextLine().next().overrideTopInset(UIUtil.LARGE_VGAP))
-      add(UiUtils.createLink("Wikipedia - List of Java bytecode instructions", "https://en.wikipedia.org/wiki/List_of_Java_bytecode_instructions"), bag.next().fillCellHorizontally().weightx(1.0).overrideLeftInset(UIUtil.DEFAULT_HGAP / 2).overrideTopInset(UIUtil.LARGE_VGAP))
+      add(
+        UiUtils.createLink("Wikipedia - List of Java bytecode instructions", "https://en.wikipedia.org/wiki/List_of_Java_bytecode_instructions"),
+        bag.next().fillCellHorizontally().weightx(1.0).overrideLeftInset(UIUtil.DEFAULT_HGAP / 2).overrideTopInset(UIUtil.LARGE_VGAP)
+      )
     }, Dimension(800, 500), project, IdeModalityType.MODELESS)
   }
 
   // -- Private Methods --------------------------------------------------------------------------------------------- //
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  private class InstructionsTable : JBTable(InstructionsTableModel()) {
+  private class InstructionsTable : JBTable(InstructionsTableModel()), DataProvider {
 
     init {
       autoResizeMode = JTable.AUTO_RESIZE_OFF
@@ -77,6 +88,18 @@ class InstructionsOverviewTool : ByteCodeTool("Instructions Overview") {
       rowSorter = TableRowSorter(model).apply {
         setComparator(1, OPCODE_COMPARATOR)
       }
+
+      addMouseListener(UiUtils.Table.createContextMenuMouseListener(InstructionsOverviewTool::class.java.simpleName) {
+        DefaultActionGroup().apply {
+          add(CopyValueAction())
+          add(ViewValueAction())
+        }
+      })
+    }
+
+    override fun getData(dataId: String): Any? = when {
+      CommonDataKeys.VALUE.`is`(dataId) -> getSingleSelectedValue(this)
+      else -> null
     }
 
     override fun prepareRenderer(renderer: TableCellRenderer, row: Int, column: Int): Component {
@@ -104,7 +127,7 @@ class InstructionsOverviewTool : ByteCodeTool("Instructions Overview") {
 
     private fun readInstructionsData(): Vector<Vector<String>> {
       val dataCsv = InstructionsOverviewTool::class.java.getResourceAsStream("/dev/turingcomplete/intellijbytecodeplugin/byte-code-instructions.csv")
-                    ?: throw IllegalStateException("snh: byte-code-instructions.csv missing")
+        ?: throw IllegalStateException("snh: byte-code-instructions.csv missing")
       return Vector(InputStreamReader(dataCsv).useLines { it.map { line -> Vector(line.split('|').toList()) }.toList() })
     }
   }