+++ CHANGELOG.md
@@ -2,12 +2,21 @@
 
 ## [Unreleased]
 ### Added
+- Automatically run code style formatting in the ASM view
+- Add Java class versions overview
 
 ### Changed
+- Compatibility improvements for IntelliJ 2023.1
+- Compatibility improvements for IntelliJ's "new UI"
+- UI improvements in the access converter
+- Change order of tabs, for better accessibility
 
 ### Removed
 
 ### Fixed
+- Fix parameters are not showing for abstract methods in the tree structure 
+- Fix misleading icon for abstract methods in the tree structure
+- Fix display of class version in the structure view if the ASM ClassNode version is represented as a negative
 
 ## [2.5.0] - 2022-10-13
 ### Added